pyproject.toml
[tool.poetry]
name = "slack-channel-manager"
version = "0.1.0"
description = "sort your slack channels in code"
authors = ["Eli Block <3347571+eliblock@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
manager = "slack_channel_manager.cli:cli"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
